Refactor: create separate action creators and reducers for all engineers and single engineer
